third_party/blink/web_tests/external/wpt/priority-hints/img-attr-named-constructor.tentative.html

<!DOCTYPE html>
<title>Priority Hints - Image element</title>
<meta name="author" title="Dominic Farolino" href="mailto:domfarolino@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<img id=img1 src=/images/green.png fetchpriority=high>
<img id=img2 src=/images/green.png fetchpriority=low>
<img id=img3 src=/images/green.png fetchpriority=auto>
<img id=img4 src=/images/green.png fetchpriority=xyz>
<img id=img5 src=/images/green.png>

<script>
  test(() => {
    assert_equals(img1.fetchpriority, "high", "high fetchpriority is a valid IDL value on the image element");
    assert_equals(img2.fetchpriority, "low", "high fetchpriority is a valid IDL value on the image element");
    assert_equals(img3.fetchpriority, "auto", "auto fetchpriority is a valid IDL value on the image element");
    assert_equals(img4.fetchpriority, "auto", "invalid fetchpriority reflects as 'auto' IDL attribute on the image element");
    assert_equals(img5.fetchpriority, "auto", "missing fetchpriority reflects as 'auto' IDL attribute on the image element");
  }, "fetchpriority attribute on <img> elements should reflect valid IDL values");

  test(() => {
    const img = new Image();
    assert_equals(img.fetchpriority, "auto");
  }, "fetchpriority of new Image() is 'auto'");
</script>
